Add a Listener type for server connections.
A Listener implements a http.Handler that upgrades its connection to a
websocket, then proxies that connection to an Accept method that the caller can
use to wait for connections to be available. Internally, the Listener maintains
a queue of pending connections for Accept, and releases the HTTP handler when
an accepted connection has closed.