listener: limit the number of backlogged connections
Instead of creating a goroutine for each pending connection, limit the number
of unaccepted (pending) connections to the buffer capacity of the queue.
Connections in excess of this are rejected without negotiating an upgrade.

The choice to reject before upgrading averts the case where a client thinks
they have a viable connection because the upgrade succeeded, only to have the
first write operation fail because the server dropped it. By rejecting first,
the client gets an ordinary HTTP repsonse indicating the error.